+++ packages/pluggableWidgets/datagrid-web/src/helpers/Column.tsx
@@ -2,7 +2,7 @@ import { isAvailable } from "@mendix/widget-plugin-platform/framework/is-availab
 import { ObjectItem } from "mendix";
 import { createElement, ReactElement } from "react";
 import { AlignmentEnum, ColumnsType, HidableEnum, WidthEnum } from "../../typings/DatagridProps";
-import { GridColumn } from "./typings/GridColumn";
+import { GridColumn } from "../typings/GridColumn";
 
 export class Column implements GridColumn {
     alignment: AlignmentEnum;
