--- packages/pluggableWidgets/datagrid-web/typings/Column.d.ts
@@ -1,3 +1,0 @@
-import { ColumnsType, ColumnsPreviewType } from "./DatagridProps";
-
-export type Column = ColumnsType | ColumnsPreviewType;
